check if ancestry dir exists before plowing thru it

+++ cells.pm
@@ -25,12 +25,14 @@ sub delete_temp_files{
     delete_unmentioned_pidfiles_in_dir(working_pids_dir(), [@active_pids] );
     delete_unmentioned_pidfiles_in_dir(current_tasks_dir(), [@active_pids] );
 
-    my @ancestry_dirs = get_directory_contents(ancestry_dir());
-    for my $ancestry_dir(@ancestry_dirs){
-        delete_unmentioned_pidfiles_in_dir($ancestry_dir, [@active_pids]);
-        my @leftovers = get_directory_contents($ancestry_dir);
-        unless(@leftovers){
-            rmdir $ancestry_dir or die "can't delete $ancestry_dir: $!";
+    if(-e ancestry_dir()){
+        my @ancestry_dirs = get_directory_contents(ancestry_dir());
+        for my $ancestry_dir(@ancestry_dirs){
+            delete_unmentioned_pidfiles_in_dir($ancestry_dir, [@active_pids]);
+            my @leftovers = get_directory_contents($ancestry_dir);
+            unless(@leftovers){
+                rmdir $ancestry_dir or die "can't delete $ancestry_dir: $!";
+            }
         }
     }
 }
